[Refactor] Move git_provider to repour/lib/scm
This is part of better organizing repour files
